test(scan): split `app_scan.test.ts` apart
Jest can run test files in parallel, but a single file always runs serially. This file was taking over 50s to run on my dev machine and commonly over 90s in CI. This commit splits it into multiple files by test subject, which should help speed up test runs.
