+++ apps/scan/backend/src/app_scan.test.ts
@@ -312,118 +312,6 @@ test('invalid ballot rejected', async () => {
   checkLogs(logger);
 });
 
-test('bmd ballot is rejected when scanned for wrong precinct', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, { precinctId: '22', mockAuth });
-  // Ballot should be rejected when configured for the wrong precinct
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'InvalidSheet',
-    reason: 'invalid_precinct',
-  };
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, {
-    state: 'rejecting',
-    interpretation,
-  });
-  await waitForStatus(apiClient, {
-    state: 'rejected',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('bmd ballot is accepted if precinct is set for the right precinct', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, { precinctId: '23', mockAuth });
-  // Configure for the proper precinct and verify the ballot scans
-
-  const validInterpretation: SheetInterpretation = {
-    type: 'ValidSheet',
-  };
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, {
-    state: 'ready_to_accept',
-    interpretation: validInterpretation,
-  });
-});
-
-test('hmpb ballot is rejected when scanned for wrong precinct', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, {
-    addTemplates: true,
-    precinctId: '22',
-    mockAuth,
-  });
-  // Ballot should be rejected when configured for the wrong precinct
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'InvalidSheet',
-    reason: 'invalid_precinct',
-  };
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, {
-    state: 'rejecting',
-    interpretation,
-  });
-  await waitForStatus(apiClient, {
-    state: 'rejected',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('hmpb ballot is accepted if precinct is set for the right precinct', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, {
-    addTemplates: true,
-    precinctId: '21',
-    mockAuth,
-  });
-  // Configure for the proper precinct and verify the ballot scans
-
-  const validInterpretation: SheetInterpretation = {
-    type: 'ValidSheet',
-  };
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, {
-    state: 'ready_to_accept',
-    interpretation: validInterpretation,
-  });
-});
-
 test('blank sheet ballot rejected', async () => {
   const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
   await configureApp(apiClient, mockUsb, { mockAuth });
@@ -449,602 +337,6 @@ test('blank sheet ballot rejected', async () => {
 
   (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
   await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('scanner powered off while waiting for paper', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  mockPlustek.simulatePowerOff();
-  await waitForStatus(apiClient, { state: 'disconnected' });
-
-  mockPlustek.simulatePowerOn();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('scanner powered off while scanning', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  mockPlustek.simulatePowerOff();
-  await waitForStatus(apiClient, { state: 'disconnected' });
-
-  mockPlustek.simulatePowerOn('jam');
-  await waitForStatus(apiClient, { state: 'jammed' });
-});
-
-test('scanner powered off while accepting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'ValidSheet',
-  };
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'ready_to_accept', interpretation });
-  mockPlustek.simulatePowerOff();
-  await apiClient.acceptBallot();
-  await waitForStatus(apiClient, { state: 'disconnected' });
-
-  mockPlustek.simulatePowerOn('ready_to_eject');
-  await waitForStatus(apiClient, {
-    state: 'rejecting',
-    error: 'paper_in_back_after_reconnect',
-  });
-  await waitForStatus(apiClient, {
-    state: 'rejected',
-    error: 'paper_in_back_after_reconnect',
-  });
-});
-
-test('scanner powered off after accepting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'ValidSheet',
-  };
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'ready_to_accept', interpretation });
-  await apiClient.acceptBallot();
-  await waitForStatus(apiClient, {
-    state: 'accepting',
-    interpretation,
-  });
-  await waitForStatus(apiClient, {
-    state: 'accepted',
-    interpretation,
-    ballotsCounted: 1,
-  });
-
-  mockPlustek.simulatePowerOff();
-  await waitForStatus(apiClient, {
-    state: 'disconnected',
-    ballotsCounted: 1,
-  });
-
-  mockPlustek.simulatePowerOn('no_paper');
-  await waitForStatus(apiClient, { state: 'no_paper', ballotsCounted: 1 });
-});
-
-test('scanner powered off while rejecting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.wrongElection)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'InvalidSheet',
-    reason: 'invalid_election_hash',
-  };
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, {
-    state: 'rejecting',
-    interpretation,
-  });
-
-  mockPlustek.simulatePowerOff();
-  await waitForStatus(apiClient, { state: 'disconnected' });
-
-  mockPlustek.simulatePowerOn('jam');
-  await waitForStatus(apiClient, { state: 'jammed' });
-});
-
-test('scanner powered off while returning', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.unmarkedHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation = needsReviewInterpretation;
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'needs_review', interpretation });
-
-  await apiClient.returnBallot();
-  await waitForStatus(apiClient, {
-    state: 'returning',
-    interpretation,
-  });
-
-  mockPlustek.simulatePowerOff();
-  await waitForStatus(apiClient, { state: 'disconnected' });
-
-  mockPlustek.simulatePowerOn('jam');
-  await waitForStatus(apiClient, { state: 'jammed' });
-});
-
-test('scanner powered off after returning', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.unmarkedHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation = needsReviewInterpretation;
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'needs_review', interpretation });
-
-  await apiClient.returnBallot();
-  await waitForStatus(apiClient, {
-    state: 'returning',
-    interpretation,
-  });
-  await waitForStatus(apiClient, {
-    state: 'returned',
-    interpretation,
-  });
-
-  mockPlustek.simulatePowerOff();
-  await waitForStatus(apiClient, { state: 'disconnected' });
-
-  mockPlustek.simulatePowerOn('ready_to_scan');
-  await waitForStatus(apiClient, {
-    state: 'rejected',
-    error: 'paper_in_front_after_reconnect',
-  });
-});
-
-test('insert second ballot while first ballot is scanning', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp({
-    delays: {},
-    mockPlustekOptions: { passthroughDuration: 500 },
-  });
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await expectStatus(apiClient, { state: 'both_sides_have_paper' });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, {
-    state: 'rejecting',
-    error: 'both_sides_have_paper',
-  });
-  await waitForStatus(apiClient, {
-    state: 'rejected',
-    error: 'both_sides_have_paper',
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('insert second ballot before first ballot accept', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'ValidSheet',
-  };
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'ready_to_accept', interpretation });
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await apiClient.acceptBallot();
-
-  await waitForStatus(apiClient, {
-    state: 'both_sides_have_paper',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_accept', interpretation });
-  await apiClient.acceptBallot();
-  await expectStatus(apiClient, { state: 'accepting', interpretation });
-  await waitForStatus(apiClient, {
-    state: 'accepted',
-    interpretation,
-    ballotsCounted: 1,
-  });
-});
-
-test('insert second ballot while first ballot is accepting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp({
-      delays: {
-        DELAY_ACCEPTED_READY_FOR_NEXT_BALLOT: 1000,
-        DELAY_ACCEPTED_RESET_TO_NO_PAPER: 2000,
-      },
-    });
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'ValidSheet',
-  };
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'ready_to_accept', interpretation });
-  await apiClient.acceptBallot();
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-
-  await waitForStatus(apiClient, {
-    state: 'accepted',
-    interpretation,
-    ballotsCounted: 1,
-  });
-  await waitForStatus(apiClient, {
-    state: 'ready_to_scan',
-    ballotsCounted: 1,
-  });
-});
-
-test('insert second ballot while first ballot needs review', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.unmarkedHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation = needsReviewInterpretation;
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'needs_review', interpretation });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.unmarkedHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, {
-    state: 'both_sides_have_paper',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'needs_review', interpretation });
-
-  await apiClient.acceptBallot();
-  await waitForStatus(apiClient, {
-    state: 'accepted',
-    interpretation,
-    ballotsCounted: 1,
-  });
-});
-
-test('insert second ballot while first ballot is rejecting', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp({
-      delays: {},
-      mockPlustekOptions: { passthroughDuration: 500 },
-    });
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.wrongElection)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'InvalidSheet',
-    reason: 'invalid_election_hash',
-  };
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, {
-    state: 'rejecting',
-    interpretation,
-  });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.wrongElection)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, {
-    state: 'both_sides_have_paper',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, {
-    state: 'rejecting',
-    interpretation,
-  });
-  await waitForStatus(apiClient, {
-    state: 'rejected',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('insert second ballot while first ballot is returning', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp({
-      delays: {},
-      mockPlustekOptions: { passthroughDuration: 500 },
-    });
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.unmarkedHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation = needsReviewInterpretation;
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'needs_review', interpretation });
-
-  await apiClient.returnBallot();
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.unmarkedHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, {
-    state: 'both_sides_have_paper',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, {
-    state: 'needs_review',
-    interpretation,
-  });
-  await apiClient.returnBallot();
-  await waitForStatus(apiClient, {
-    state: 'returned',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('jam on scan', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp({
-    delays: {
-      DELAY_RECONNECT_ON_UNEXPECTED_ERROR: 500,
-    },
-  });
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  mockPlustek.simulateJamOnNextOperation();
-  await apiClient.scanBallot();
-  await waitForStatus(apiClient, {
-    state: 'recovering_from_error',
-    error: 'plustek_error',
-  });
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('jam on accept', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp({
-      delays: {
-        DELAY_ACCEPTING_TIMEOUT: 500,
-      },
-    });
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'ValidSheet',
-  };
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await waitForStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'ready_to_accept', interpretation });
-
-  mockPlustek.simulateJamOnNextOperation();
-  await apiClient.acceptBallot();
-  await waitForStatus(apiClient, { state: 'accepting', interpretation });
-  // The paper can't get permanently jammed on accept - it just stays held in
-  // the back and we can reject at that point
-  await waitForStatus(apiClient, {
-    state: 'rejecting',
-    interpretation,
-    error: 'paper_in_back_after_accept',
-  });
-  await waitForStatus(apiClient, {
-    state: 'rejected',
-    error: 'paper_in_back_after_accept',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('jam on return', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.unmarkedHmpb)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation = needsReviewInterpretation;
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  await waitForStatus(apiClient, { state: 'needs_review', interpretation });
-
-  mockPlustek.simulateJamOnNextOperation();
-  await apiClient.returnBallot();
-  await waitForStatus(apiClient, {
-    state: 'jammed',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('jam on reject', async () => {
-  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
-    await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (
-    await mockPlustek.simulateLoadSheet(ballotImages.wrongElection)
-  ).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const interpretation: SheetInterpretation = {
-    type: 'InvalidSheet',
-    reason: 'invalid_election_hash',
-  };
-  mockInterpretation(interpreter, interpretation);
-
-  await apiClient.scanBallot();
-  await expectStatus(apiClient, { state: 'scanning' });
-  mockPlustek.simulateJamOnNextOperation();
-  await waitForStatus(apiClient, {
-    state: 'jammed',
-    interpretation,
-  });
-
-  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'no_paper' });
-});
-
-test('calibrate', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (await mockPlustek.simulateLoadSheet(ballotImages.blankSheet)).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  const calibratePromise = apiClient.calibrate();
-  await waitForStatus(apiClient, { state: 'calibrating' });
-  expect(await calibratePromise).toEqual(true);
-  await expectStatus(apiClient, { state: 'no_paper' });
-});
-
-test('calibrate not supported', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (await mockPlustek.simulateLoadSheet(ballotImages.blankSheet)).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  mockPlustek.simulateCalibrateNotSupported();
-  const calibrateResult = await apiClient.calibrate();
-  expect(calibrateResult).toEqual(false);
-});
-
-test('jam on calibrate', async () => {
-  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp();
-  await configureApp(apiClient, mockUsb, { mockAuth });
-
-  (await mockPlustek.simulateLoadSheet(ballotImages.blankSheet)).unsafeUnwrap();
-  await waitForStatus(apiClient, { state: 'ready_to_scan' });
-
-  mockPlustek.simulateJamOnNextOperation();
-  expect(await apiClient.calibrate()).toEqual(false);
-  await expectStatus(apiClient, { state: 'jammed' });
 });
 
 test('scan fails and retries', async () => {
