+++ apps/scan/backend/src/app_scan_jam.test.ts
@@ -1,0 +1,180 @@
+import { Logger } from '@votingworks/logging';
+import { AdjudicationReason } from '@votingworks/types';
+import {
+  ballotImages,
+  configureApp,
+  createApp,
+  expectStatus,
+  mockInterpretation,
+  waitForStatus,
+} from '../test/helpers/app_helpers';
+import { SheetInterpretation } from './types';
+
+jest.setTimeout(20_000);
+jest.mock('@votingworks/ballot-encoder', () => {
+  return {
+    ...jest.requireActual('@votingworks/ballot-encoder'),
+    // to allow changing election definitions without changing the image fixtures
+    // TODO: generate image fixtures from election definitions more easily
+    // this election hash is for the famous names image fixtures
+    sliceElectionHash: () => 'da81438d51136692b43c',
+  };
+});
+
+// Basic checks for logging. We don't try to be exhaustive here because paper
+// status polling can be a bit non-deterministic, so logs can vary between runs.
+export function checkLogs(logger: Logger): void {
+  // Make sure we got a transition
+  expect(logger.log).toHaveBeenCalledWith(
+    'scanner-state-machine-transition',
+    'system',
+    { message: 'Transitioned to: "checking_initial_paper_status"' },
+    expect.any(Function)
+  );
+  // Make sure we got an event
+  expect(logger.log).toHaveBeenCalledWith(
+    'scanner-state-machine-event',
+    'system',
+    { message: 'Event: SCANNER_NO_PAPER' },
+    expect.any(Function)
+  );
+  // Make sure we got a context update. And make sure we didn't log the votes in
+  // the interpretation, just the type, to protect voter privacy.
+  expect(logger.log).toHaveBeenCalledWith(
+    'scanner-state-machine-transition',
+    'system',
+    {
+      message: 'Context updated',
+      changedFields: expect.stringMatching(
+        /{"interpretation":"(ValidSheet|InvalidSheet|NeedsReviewSheet)"}/
+      ),
+    },
+    expect.any(Function)
+  );
+}
+
+const needsReviewInterpretation: SheetInterpretation = {
+  type: 'NeedsReviewSheet',
+  reasons: [{ type: AdjudicationReason.BlankBallot }],
+};
+
+test('jam on scan', async () => {
+  const { apiClient, mockPlustek, mockUsb, mockAuth } = await createApp({
+    delays: {
+      DELAY_RECONNECT_ON_UNEXPECTED_ERROR: 500,
+    },
+  });
+  await configureApp(apiClient, mockUsb, { mockAuth });
+
+  (
+    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
+  ).unsafeUnwrap();
+  await waitForStatus(apiClient, { state: 'ready_to_scan' });
+
+  mockPlustek.simulateJamOnNextOperation();
+  await apiClient.scanBallot();
+  await waitForStatus(apiClient, {
+    state: 'recovering_from_error',
+    error: 'plustek_error',
+  });
+  await waitForStatus(apiClient, { state: 'no_paper' });
+});
+
+test('jam on accept', async () => {
+  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
+    await createApp({
+      delays: {
+        DELAY_ACCEPTING_TIMEOUT: 500,
+      },
+    });
+  await configureApp(apiClient, mockUsb, { mockAuth });
+
+  (
+    await mockPlustek.simulateLoadSheet(ballotImages.completeBmd)
+  ).unsafeUnwrap();
+  await waitForStatus(apiClient, { state: 'ready_to_scan' });
+
+  const interpretation: SheetInterpretation = {
+    type: 'ValidSheet',
+  };
+  mockInterpretation(interpreter, interpretation);
+
+  await apiClient.scanBallot();
+  await waitForStatus(apiClient, { state: 'scanning' });
+  await waitForStatus(apiClient, { state: 'ready_to_accept', interpretation });
+
+  mockPlustek.simulateJamOnNextOperation();
+  await apiClient.acceptBallot();
+  await waitForStatus(apiClient, { state: 'accepting', interpretation });
+  // The paper can't get permanently jammed on accept - it just stays held in
+  // the back and we can reject at that point
+  await waitForStatus(apiClient, {
+    state: 'rejecting',
+    interpretation,
+    error: 'paper_in_back_after_accept',
+  });
+  await waitForStatus(apiClient, {
+    state: 'rejected',
+    error: 'paper_in_back_after_accept',
+    interpretation,
+  });
+
+  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
+  await waitForStatus(apiClient, { state: 'no_paper' });
+});
+
+test('jam on return', async () => {
+  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
+    await createApp();
+  await configureApp(apiClient, mockUsb, { mockAuth });
+
+  (
+    await mockPlustek.simulateLoadSheet(ballotImages.unmarkedHmpb)
+  ).unsafeUnwrap();
+  await waitForStatus(apiClient, { state: 'ready_to_scan' });
+
+  const interpretation = needsReviewInterpretation;
+  mockInterpretation(interpreter, interpretation);
+
+  await apiClient.scanBallot();
+  await expectStatus(apiClient, { state: 'scanning' });
+  await waitForStatus(apiClient, { state: 'needs_review', interpretation });
+
+  mockPlustek.simulateJamOnNextOperation();
+  await apiClient.returnBallot();
+  await waitForStatus(apiClient, {
+    state: 'jammed',
+    interpretation,
+  });
+
+  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
+  await waitForStatus(apiClient, { state: 'no_paper' });
+});
+
+test('jam on reject', async () => {
+  const { apiClient, mockPlustek, interpreter, mockUsb, mockAuth } =
+    await createApp();
+  await configureApp(apiClient, mockUsb, { mockAuth });
+
+  (
+    await mockPlustek.simulateLoadSheet(ballotImages.wrongElection)
+  ).unsafeUnwrap();
+  await waitForStatus(apiClient, { state: 'ready_to_scan' });
+
+  const interpretation: SheetInterpretation = {
+    type: 'InvalidSheet',
+    reason: 'invalid_election_hash',
+  };
+  mockInterpretation(interpreter, interpretation);
+
+  await apiClient.scanBallot();
+  await expectStatus(apiClient, { state: 'scanning' });
+  mockPlustek.simulateJamOnNextOperation();
+  await waitForStatus(apiClient, {
+    state: 'jammed',
+    interpretation,
+  });
+
+  (await mockPlustek.simulateRemoveSheet()).unsafeUnwrap();
+  await waitForStatus(apiClient, { state: 'no_paper' });
+});
